requirements.txt
gymnasium
playwright==1.30.0
Pillow
evaluate
openai==0.27.0
types-tqdm
tiktoken
aiolimiter
beartype==0.12.0
flask
tornado
matplotlib==3.2.1
nltk
google-generativeai
text-generation
scikit-learn
transformers==4.33.2
sentencepiece
accelerate
pyvis
sentence-transformers
zeno_client
python-dotenv